requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/production.txt requirements/production.in
#
apscheduler==3.6.3
    # via python-telegram-bot
bcrypt==3.2.0
    # via -r requirements/production.in
certifi==2020.12.5
    # via python-telegram-bot
cffi==1.14.5
    # via bcrypt
click==7.1.2
    # via flask
cycler==0.10.0
    # via matplotlib
flask==1.1.2
    # via -r requirements/production.in
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
kiwisolver==1.3.1
    # via matplotlib
markupsafe==1.1.1
    # via jinja2
matplotlib==3.4.1
    # via -r requirements/production.in
numpy==1.20.2
    # via
    #   -r requirements/production.in
    #   matplotlib
    #   torch
    #   torchvision
pillow==8.2.0
    # via
    #   matplotlib
    #   torchvision
pycparser==2.20
    # via cffi
pymongo==3.11.3
    # via -r requirements/production.in
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via matplotlib
python-dotenv==0.17.0
    # via -r requirements/production.in
python-telegram-bot==13.4.1
    # via -r requirements/production.in
pytz==2021.1
    # via
    #   apscheduler
    #   python-telegram-bot
six==1.15.0
    # via
    #   apscheduler
    #   bcrypt
    #   cycler
    #   python-dateutil
torch==1.8.1
    # via
    #   -r requirements/production.in
    #   torchvision
torchvision==0.9.1
    # via -r requirements/production.in
tornado==6.1
    # via python-telegram-bot
typing-extensions==3.7.4.3
    # via torch
tzlocal==2.1
    # via apscheduler
werkzeug==1.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
